move analyst data-specific constants out of aws; cleanup
